% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide-class.R
\name{new_peptide}
\alias{new_peptide}
\title{Internal constructor to create \code{peptr_peptide} type}
\usage{
new_peptide(sequence = character(), ptms = list(), position = peptr_position())
}
\description{
Asserts that...
}
\keyword{internal}
