% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm-class.R
\name{new_ptm}
\alias{new_ptm}
\title{Internal constructor to create \code{peptr_ptm} type}
\usage{
new_ptm(position = peptr_position(), name = "unknown")
}
\description{
Asserts that \code{position} is a vector of integers, and that \code{name} is a string.
}
\keyword{internal}
