% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_position-class.R
\name{new_value_position}
\alias{new_value_position}
\title{Internal constructor to create \code{peptr_value_position} type}
\usage{
new_value_position(
  value = double(),
  position = integer(),
  range = double(),
  name = "unknown"
)
}
\description{
Asserts that \code{value} is a double value, \code{position} is an integer vector,
and that \code{name} is a string.
}
\keyword{internal}
