% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptr-package.R
\docType{package}
\name{peptr-package}
\alias{peptr}
\alias{peptr-package}
\title{peptr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jean Manguy \email{jean@manguy.eu} (\href{https://orcid.org/0000-0002-8427-007X}{ORCID})

}
\keyword{internal}
