% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm-class.R
\name{peptr_get_ptm_name}
\alias{peptr_get_ptm_name}
\title{Access the name attribute of a \code{peptr_ptm} object.}
\usage{
peptr_get_ptm_name(x)
}
\description{
Access the name attribute of a \code{peptr_ptm} object.
}
\keyword{internal}
