% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm-class.R
\name{peptr_is_ptm}
\alias{peptr_is_ptm}
\title{Test if an object is of class \code{peptr_ptm}}
\usage{
peptr_is_ptm(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if object is of class \code{peptr_ptm} and \code{FALSE} if it is not.
}
\description{
Test if an object is of class \code{peptr_ptm}
}
\examples{
x <- peptr_ptm(c(5, 3, 8))
y <- c(5, 3, 8)

peptr_is_ptm(x)
peptr_is_ptm(y)
}
