% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-class.R
\name{peptr_position}
\alias{peptr_position}
\title{A decimalized class for pounds, shillings, and pence values}
\usage{
peptr_position(x = integer())
}
\arguments{
\item{x}{An integer vectore representing positions}
}
\description{
A decimalized class for pounds, shillings, and pence values
}
