% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm-class.R
\name{peptr_ptm}
\alias{peptr_ptm}
\title{A class for post-translational modifications that occurs on one residue}
\usage{
peptr_ptm(position = integer(), name = "unknown")
}
\arguments{
\item{position}{An integer vector of the modified residues positions}

\item{name}{A string to identify the PTM (e.g. Phosphorylation, Default: "unknown")}
}
\description{
A class for post-translational modifications that occurs on one residue
}
\examples{
peptr_ptm(c(10L, 22L, 58L, 125L), "phosphorylation")
peptr_ptm(c(58L, 132L, 24L), "O-glycosylation")
}
