% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide-class.R, R/ptm2-class.R
\name{peptr_ptm2}
\alias{peptr_ptm2}
\title{A class for peptidic sequences}
\usage{
peptr_ptm2(position_1 = integer(), position_2 = integer(), name = "unknown")

peptr_ptm2(position_1 = integer(), position_2 = integer(), name = "unknown")
}
\arguments{
\item{position_1}{An integer vector of the modified residues start positions}

\item{position_2}{An integer vector of the modified residues end positions}

\item{name}{A string to identify the PTM (e.g. disulphide bond, Default: "unknown")}
}
\description{
A class for peptidic sequences

A class for post-translational modifications that occurs between 2 residues
}
\examples{
peptr_ptm2(c(10L, 22L, 58L, 125L), c(15L, 137L, 62L, 60L), "disulphide bond")
peptr_ptm2(c(10L, 22L, 58L, 125L), c(15L, 137L, 62L, 60L), "disulphide bond")
}
