% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_position-class.R
\name{peptr_value_position}
\alias{peptr_value_position}
\title{A class for position specific variable}
\usage{
peptr_value_position(
  value = double(),
  position = peptr_position(),
  range = double(),
  name = "unknown"
)
}
\arguments{
\item{value}{A double vector of the residue value}

\item{position}{An integer vector of the residue position}

\item{range}{A vector of 2 doubles to scale \code{value}}

\item{name}{A string to identify the variable (e.g. disorder, Default: "unknown")}
}
\description{
A class for position specific variable
}
\examples{
peptr_value_position(c(0.5, 0.3, 0.7), c(1, 2, 3), range = c(0, 1), name = "hydrophobicity")
}
