% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptr-package.R, R/position-class.R
\name{vctrs-compat}
\alias{vctrs-compat}
\alias{vec_cast.peptr_position}
\title{vctrs compatibility functions for maths, casting, coercion, and comparison}
\usage{
\method{vec_cast}{peptr_position}(x, to, ...)
}
\arguments{
\item{x, y}{Vectors.}

\item{to}{Type to cast to.}

\item{...}{Arguments passed on to further methods.}

\item{op}{Arithmetic operation,}

\item{.fn}{Mathematical function.}

\item{.x}{Vector}
}
\description{
vctrs compatibility functions for maths, casting, coercion, and comparison
}
