% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-class.R
\name{peptr_position}
\alias{peptr_position}
\title{A for positions on peptidic chains}
\usage{
peptr_position(x = integer())
}
\arguments{
\item{x}{An integer vector representing positions}
}
\description{
A for positions on peptidic chains
}
\examples{
peptr_position(1:10)
}
