% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_position-class.R
\name{peptr_value_position}
\alias{peptr_value_position}
\title{A class for position specific variable}
\usage{
peptr_value_position(
  value = double(),
  position = integer(),
  range = double(),
  name = "unknown"
)
}
\arguments{
\item{value}{A double vector of the residue value}

\item{position}{An integer vector of the residue position}

\item{name}{A string to identify the variable (e.g. disorder, Default: "unknown")}
}
\description{
A class for position specific variable
}
\examples{
peptr_value_position(0.5, 1, c(0, 1))
}
